+++ src/exception/WordException.js
@@ -1,0 +1,10 @@
+export default class WordException {
+  constructor() {
+    this.name = 'WordException';
+    this.message = 'Try another word!';
+  }
+
+  toString() {
+    return `${this.name}: ${this.message}`;
+  }
+}