+++ src/level/InsaneLevel.js
@@ -1,0 +1,9 @@
+import Level from './Level';
+
+export default class InsaneLevel extends Level {
+  constructor() {
+    super();
+    this.name = 'InsaneLevelBehavior';
+    this.level = 'insane';
+  }
+}