+++ src/rule/Rule.js
@@ -1,0 +1,7 @@
+const Rule = {
+  ALLOWED_INITIAL: 0,
+  ALLOWED_EXTRA: 1,
+  MANNER: 2,
+};
+
+export default Rule;